refactor: 抽离http请求方法
