+++ adapter/provider/vehicle.go
@@ -2,16 +2,12 @@ package provider
 
 import (
 	"context"
-	"github.com/Dreamacro/clash/listener/inner"
+	netHttp "github.com/Dreamacro/clash/component/http"
+	types "github.com/Dreamacro/clash/constant/provider"
 	"io"
-	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"time"
-
-	netHttp "github.com/Dreamacro/clash/common/net"
-	types "github.com/Dreamacro/clash/constant/provider"
 )
 
 type FileVehicle struct {
@@ -50,52 +46,16 @@ func (h *HTTPVehicle) Path() string {
 func (h *HTTPVehicle) Read() ([]byte, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*20)
 	defer cancel()
-
-	uri, err := url.Parse(h.url)
+	resp, err := netHttp.HttpRequest(ctx, h.url, http.MethodGet, nil, nil)
 	if err != nil {
 		return nil, err
 	}
 
-	req, err := http.NewRequest(http.MethodGet, uri.String(), nil)
-	req.Header.Set("User-Agent", netHttp.UA)
-
-	if err != nil {
-		return nil, err
-	}
-
-	if user := uri.User; user != nil {
-		password, _ := user.Password()
-		req.SetBasicAuth(user.Username(), password)
-	}
-
-	req = req.WithContext(ctx)
-
-	transport := &http.Transport{
-		// from http.DefaultTransport
-		MaxIdleConns:          100,
-		IdleConnTimeout:       30 * time.Second,
-		TLSHandshakeTimeout:   10 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-		DialContext: func(ctx context.Context, network, address string) (net.Conn, error) {
-			conn := inner.HandleTcp(address, uri.Hostname())
-			return conn, nil
-		},
-	}
-
-	client := http.Client{Transport: transport}
-	resp, err := client.Do(req)
-	if err != nil {
-		if err != nil {
-			return nil, err
-		}
-	}
 	defer resp.Body.Close()
-
 	buf, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return nil, err
 	}
-
 	return buf, nil
 }
