--- common/net/http.go
@@ -1,5 +1,0 @@
-package net
-
-const (
-	UA = "Clash"
-)
